separated the sample into 2 more targetted samples (XML vs. programmatic) like it is done for the clustered sample
